requirements/base.txt
psycopg2==2.5.4
mezzanine==3.1.10
South
Pillow==2.6.1
django-compressor==1.4
django-model-utils==2.2
django-vanilla-views==1.0.3
loremipsum